+++ src/app/vaccineinfo/page.tsx
@@ -1,7 +1,7 @@
 import AgeFilter from "@/components/vaccineinfo/AgeFilter";
 import AgeGroup from "@/components/vaccineinfo/AgeGroup";
 import VaccineList from "@/components/vaccineinfo/VaccineList";
-import { getVaccineList } from "@/utils/vaccineInfo/server-action";
+import { getVaccineList } from "@/utils/supabase/server-action";
 
 const VaccineInfoPage = async () => {
   const vaccineList = await getVaccineList();
